+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Upstream merge
 - Fix PervEden search
 - Add ability to use high-quality thumbnails on nhentai
-- Enable PervEden link importing
+- Enable link importing for all NSFW sources
 - Fix back button in library search
 - Add HentaiCafe source
+- Add Tsumino source